메세지출력부분 파일 생성하고 사다리넓이 수정
메세지출력하는 부분을 파일로 분리
사다리 넓이를 최대입력글자수인 5개로 수정
